Django/requirements.txt
django==3.1.1
seaborn==0.11.0
django-cors-headers==3.5.0
fastcluster==1.1.26
distinctipy==1.1.5
mplcairo==0.3
PyGObject==3.38
gunicorn==20.0.4
python-decouple==3.3
django-cprofile-middleware==1.0.5
plotly==4.14.1
uWSGI==2.0.20
pybedtools==0.8.2
django-debug-toolbar==3.2.2
requests==2.22.0
django-debug-toolbar-force==0.1.8